requirements.txt
transformers==4.50.3
streamlit==1.44.1
chromadb==0.6.3
requests==2.32.3
torch==2.6.0
python-dotenv==1.1.0
opencv-python==4.11.0.86